azure_utils/speech_to_text.py

import azure.cognitiveservices.speech as speechsdk

def transcribe_audio(audio_path, speech_key, speech_region):
    speech_config = speechsdk.SpeechConfig(subscription=speech_key, region=speech_region)
    audio_config = speechsdk.AudioConfig(filename=audio_path)
    recognizer = speechsdk.SpeechRecognizer(speech_config=speech_config, audio_config=audio_config)
    result = recognizer.recognize_once()
    return result.text if result.reason == speechsdk.ResultReason.RecognizedSpeech else "Could not recognize speech"
