requirements.txt
streamlit==1.37.1
# streamlit-webrtc==0.45.0
streamlit-webrtc==0.61.0
numpy==1.24.3
azure-cognitiveservices-speech==1.33.0
azure-ai-language-conversations==1.1.0
requests==2.31.0
